Move InputCfg into own file
