+++ examples/hc_num_converter.rs
@@ -5,8 +5,8 @@
 use std::sync::Arc;
 use std::time::Duration;
 
+use librunen::rnn::common::input_config::InputCfg;
 use librunen::rnn::layouts::network::Network;
-use librunen::rnn::neural::dendrite::InputCfg;
 use tokio::task;
 use tokio::time::sleep;
