+++ src/rnn/common.rs
@@ -2,3 +2,5 @@ pub mod rnn_error;
 pub mod utils;
 
 pub mod spec_type;
+
+pub mod input_config;
