+++ src/rnn/layouts/network.rs
@@ -9,10 +9,10 @@ use tokio::sync::broadcast::{self, Receiver, Sender};
 use tokio::sync::RwLock;
 use tokio::task::JoinHandle;
 
+use crate::rnn::common::input_config::InputCfg;
 use crate::rnn::common::rnn_error::RnnError;
 use crate::rnn::common::spec_type::SpecificationType;
 use crate::rnn::common::utils::gen_id_by_spec_type;
-use crate::rnn::neural::dendrite::InputCfg;
 use crate::rnn::neural::neuron::{Neuron, NeuronConfig};
 
 static mut ID_COUNTER: AtomicUsize = AtomicUsize::new(0_usize);
